% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Canon_sampling}
\alias{Canon_sampling}
\title{Canon without off target option}
\usage{
Canon_sampling(
  xin,
  yin,
  gRNAin,
  Gibbsnumberin,
  burninproportion,
  initial_betain,
  initial_gammain,
  pi_beta_shape_in,
  pi_beta_scale_in,
  a_betain,
  b_betain,
  sigma_dfin,
  verbosein
)
}
\description{
Canon without off target option
}
