% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Canon_sampling_off_target}
\alias{Canon_sampling_off_target}
\title{Canon with off target effects}
\usage{
Canon_sampling_off_target(
  xin,
  yin,
  gRNAin,
  Gibbsnumberin,
  burninproportion,
  initial_betain,
  initial_gammain,
  pi_beta_shape_in,
  pi_beta_scale_in,
  a_betain,
  b_betain,
  pi_eta_shape_in,
  pi_eta_scale_in,
  a_etain,
  b_etain,
  sigma_dfin,
  verbosein
)
}
\description{
Canon with off target effects
}
